work in proress separating out the plotting functions to a separate p…

+++ src/BallTree01.jl
@@ -8,19 +8,19 @@ FIELD_NAMES = ["D", "N", "centers", "ranges", "weights",
 nfields = 10;
 
 type BallTree
-  dims::Int64                     # dimension of data
-  num_points::Int64               # of points
+  dims::Int                     # dimension of data
+  num_points::Int               # of points
   centers::Array{Float64,1}       # ball centers, dims numbers per ball
   ranges::Array{Float64,1}        # bounding box ranges, dims per ball, dist from center to one side
   weights::Array{Float64,1}       # total weight in each ball
 
-  left_child::Array{Int64,1}
-  right_child::Array{Int64,1}     # left, right children; no parent indices
-  lowest_leaf::Array{Int64,1}
-  highest_leaf::Array{Int64,1}    # lower & upper leaf indices for each ball
-  permutation::Array{Int64,1}     # point's position in the original data
+  left_child::Array{Int,1}
+  right_child::Array{Int,1}     # left, right children; no parent indices
+  lowest_leaf::Array{Int,1}
+  highest_leaf::Array{Int,1}    # lower & upper leaf indices for each ball
+  permutation::Array{Int,1}     # point's position in the original data
 
-  next::Int64                     # internal var for placing the non-leaf nodes
+  next::Int                     # internal var for placing the non-leaf nodes
 
   swapHandle::Function
   calcStatsHandle::Function
@@ -31,49 +31,49 @@ end
 root() = 1
 Ndim(bt::BallTree) = bt.dims
 Npts(bt::BallTree) = bt.num_points
-Npts(bt::BallTree, i::Int64) = bt.highest_leaf[i]-bt.lowest_leaf[i]+1
+Npts(bt::BallTree, i::Int) = bt.highest_leaf[i]-bt.lowest_leaf[i]+1
 
 
 ## todo make sure these two are working properly -- this is the problem, we are always assigning new memory for each return
-#function center(bt::BallTree, i::Int64)
+#function center(bt::BallTree, i::Int)
 #    return bt.centers[((i-1)*bt.dims+1):end]
 #end
-center(bt::BallTree, i::Int64) = bt.centers[((i-1)*bt.dims+1):end]
-center(bt::BallTree, i::Int64, k::Int) = bt.centers[((i-1)*bt.dims+k)]
+center(bt::BallTree, i::Int) = bt.centers[((i-1)*bt.dims+1):end]
+center(bt::BallTree, i::Int, k::Int) = bt.centers[((i-1)*bt.dims+k)]
 
-#function rangeB(bt::BallTree, i::Int64)
+#function rangeB(bt::BallTree, i::Int)
 #    return bt.ranges[((i-1)*bt.dims+1):end]
 #end
-rangeB(bt::BallTree, i::Int64) = bt.ranges[((i-1)*bt.dims+1):end]
-rangeB(bt::BallTree, i::Int64, k::Int) = bt.ranges[((i-1)*bt.dims+k)]
+rangeB(bt::BallTree, i::Int) = bt.ranges[((i-1)*bt.dims+1):end]
+rangeB(bt::BallTree, i::Int, k::Int) = bt.ranges[((i-1)*bt.dims+k)]
 
-#function weight(bt::BallTree, i::Int64)
+#function weight(bt::BallTree, i::Int)
 #    return bt.weights[i]
 #end
-weight(bt::BallTree, i::Int64) = bt.weights[i]
+weight(bt::BallTree, i::Int) = bt.weights[i]
 
-#function isLeaf(bt::BallTree, ind::Int64)
+#function isLeaf(bt::BallTree, ind::Int)
 #    return ind >= bt.num_points
 #end
-isLeaf(bt::BallTree, ind::Int64) = ind >= bt.num_points
+isLeaf(bt::BallTree, ind::Int) = ind >= bt.num_points
 
-validIndex(bt::BallTree, ind::Int64) = ((0<ind) && (ind <= 2*bt.num_points))
+validIndex(bt::BallTree, ind::Int) = ((0<ind) && (ind <= 2*bt.num_points))
 
-left(bt::BallTree, i::Int64) = bt.left_child[i]
+left(bt::BallTree, i::Int) = bt.left_child[i]
 
-right(bt::BallTree, i::Int64) = bt.right_child[i]
+right(bt::BallTree, i::Int) = bt.right_child[i]
 
-leafFirst(bt::BallTree, i::Int64) = bt.lowest_leaf[i]
+leafFirst(bt::BallTree, i::Int) = bt.lowest_leaf[i]
 
-leafLast(bt::BallTree, i::Int64) = bt.highest_leaf[i]
+leafLast(bt::BallTree, i::Int) = bt.highest_leaf[i]
 
 # Convert a BallTree::index to the numeric index in the original data
-getIndexOf(bt::BallTree, i::Int64) = bt.permutation[i]
+getIndexOf(bt::BallTree, i::Int) = bt.permutation[i]
 
-function swap!(data, _i::Int64, _j::Int64)
+function swap!(data, _i::Int, _j::Int)
   return data.swapHandle(data, _i, _j)
 end
-function calcStats!(data, root::Int64)
+function calcStats!(data, root::Int)
   #@show "Fancy calcStats"
   return data.calcStatsHandle(data, root)
 end
@@ -83,7 +83,7 @@ end
 # weights, permutation, and centers, so only for swapping
 # leaves. Will not swap ranges correctly and will not swap children
 # correctly.
-function swapBall!(bt::BallTree, _i::Int64, _j::Int64)
+function swapBall!(bt::BallTree, _i::Int, _j::Int)
   #println("swapBall! -- (i,j)=$((_i,_j))")
   i = _i
   j = _j
@@ -116,7 +116,7 @@ end
 
 # Find the dimension along which the leaves between low and high
 # inclusive have the greatest variance
-function  most_spread_coord(bt::BallTree, low::Int64, high::Int64)
+function  most_spread_coord(bt::BallTree, low::Int, high::Int)
   #BallTree::index dimension, point, max_dim;
   #double mean, variance, max_variance;
   #println("most_spread_coord -- low, high = $((low, high))")
@@ -148,7 +148,7 @@ end
 # quicksort.  Partitions the leaves from low to high inclusive around
 # a random pivot in the given dimension.  Does not affect non-leaf
 # nodes, but does relabel the leaves from low to high.
-function partition!(bt::BallTree, dimension::Int64, low::Int64, high::Int64)
+function partition!(bt::BallTree, dimension::Int, low::Int, high::Int)
   pivot = low;  # not randomized, could set pivot to a random element
 
   while (low < high)
@@ -170,12 +170,12 @@ end
 # Function to partition the data into two (equal-sized or near as possible)
 #   sets, one of which is uniformly greater than the other in the given
 #   dimension.
-function select!(bt::BallTree, dimension::Int64, position::Int64, low::Int64, high::Int64)
+function select!(bt::BallTree, dimension::Int, position::Int, low::Int, high::Int)
     m = 0
     r = 0
     i = 0
   while low < high
-    r = (floor(Int64,(low + high)/2))
+    r = (floor(Int,(low + high)/2))
     swap!(bt.data, r, low)
     m = low;
     for i in (low):high
@@ -194,7 +194,7 @@ end
 
 # Calculate the statistics of level "root" based on the statistics of
 #   its left and right children.
-function calcStatsBall!(bt::BallTree, root::Int64)
+function calcStatsBall!(bt::BallTree, root::Int)
   #println("calcStatsBall! -- root=$(root)")
   Ni = 0
   NiL = 0
@@ -246,7 +246,7 @@ end
 # Given the leaves, build the rest of the tree from the top down.
 # Split the leaves along the most spread coordinate, build two balls
 # out of those, and then build a ball around those two children.
-function buildBall!(bt::BallTree, low::Int64, high::Int64, root::Int64)
+function buildBall!(bt::BallTree, low::Int, high::Int, root::Int)
   #println("buildBall! -- (low, high, root)=$((low, high, root))")
   # special case for N=1 trees
   if (low == high)
@@ -269,7 +269,7 @@ function buildBall!(bt::BallTree, low::Int64, high::Int64, root::Int64)
   # Choose the most spread coordinate to split them on, and make sure
   # there are the same number of points in each (+-1 for round off
   # error).
-  split = (floor(Int64,(low + high) / 2))
+  split = (floor(Int,(low + high) / 2))
   #@show coord, split, low, high
   select!(bt, coord, split, low, high)
 
@@ -347,8 +347,8 @@ function makeBallTree(_pointsMatrix::Array{Float64,2}, _weights::Array{Float64,1
   weights[(Np+1):end] = _weights[1:end]
 
   bt = BallTree(Nd, Np, centers, ranges, weights,
-                  ones(Int64,2*Np), ones(Int64,2*Np), ones(Int64,2*Np),
-                  ones(Int64,2*Np), zeros(Int64,2*Np),
+                  ones(Int,2*Np), ones(Int,2*Np), ones(Int,2*Np),
+                  ones(Int,2*Np), zeros(Int,2*Np),
                   0, swapBall!, calcStatsBall!, [])
   bt.data = bt
